Move typography.scss from assets to sources, because it generates locally scoped selectors.
