+++ app/src/docs/_examples/richTextView/BasicExample.module.scss
@@ -1,11 +1,5 @@
 @use 'epam-promo/assets/styles/typography-mixins.scss' as epamPromoMixins;
 
-.container {
-    a svg {
-        margin-left: 4px;
-    }
-}
-
 .typography-16 {
     @include epamPromoMixins.typography-16();
 }
