+++ app/src/docs/_props/loveship/components/typography/richTextViewDoc.module.scss
@@ -1,5 +1,15 @@
+@use 'loveship/assets/styles/scss/typography-mixins.scss' as epamLoveshipMixins;
+
 .container {
     a svg {
         margin-left: 4px;
     }
+}
+
+.typography-16 {
+    @include epamLoveshipMixins.typography-16();
+}
+
+.typography-12 {
+    @include epamLoveshipMixins.typography-12();
 }
