--- epam-promo/assets/styles/typography.scss
@@ -1,36 +1,0 @@
-@forward 'typography-mixins';
-@use 'typography-mixins' as *;
-@use 'colors' as *;
-@use 'fonts' as *;
-
-:local(.typography-header) {
-    @include typography-header();
-}
-
-:local(.typography-block) {
-    @include typography-block();
-}
-
-:local(.typography-inline) {
-    @include typography-inline();
-}
-
-:local(.typography-16) {
-    @include typography-16();
-}
-
-:local(.typography-14) {
-    @include typography-14();
-}
-
-:local(.typography-12) {
-    @include typography-12();
-}
-
-:local(.typography-promo) {
-    font-family: $font-sans;
-    color: $gray80;
-    composes: typography-header;
-    composes: typography-block;
-    composes: typography-inline;
-}
