+++ uui/assets/styles/index.scss
@@ -5,4 +5,4 @@
 @forward 'inputs';
 @forward 'layout-mixins';
 @forward 'text-size';
-@forward 'typography';
+@forward 'typography-mixins';
