+++ uui/components/typography/RichTextView.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import style from '../../assets/styles/typography.scss';
+import style from './RichTextView.module.scss';
 import { RichTextView as uuiRichTextView, RichTextViewProps } from '@epam/uui-components';
 import { withMods } from '@epam/uui-core';
